+++ src/mcmgr_internal_core_api_kw47b42.c
@@ -40,9 +40,6 @@ const mcmgr_system_info_t g_mcmgrSystem = {
 
 mcmgr_status_t mcmgr_platform_init_internal_early(mcmgr_core_t coreNum)
 {
-    /* This function is intended to be called as close to the reset entry as possible,
-       (within the startup sequence in SystemInitHook) to allow CoreUp event triggering.
-       Avoid using uninitialized data here. */
     mcmgr_status_t status = kStatus_MCMGR_Error;
 
     mcmgr_imu_remote_active_req();
@@ -64,9 +61,9 @@ mcmgr_status_t mcmgr_platform_init_internal_early(mcmgr_core_t coreNum)
 
     /* Trigger core up event here, core is starting! */
 #if (defined(MCMGR_BUILD_FOR_CORE_0))
-    status = MCMGR_TriggerEvent(kMCMGR_Core1, kMCMGR_RemoteCoreUpEvent, 0);
+    status = mcmgr_trigger_event_internal(kMCMGR_Core1, kMCMGR_RemoteCoreUpEvent, 0U, false);
 #else
-    status = MCMGR_TriggerEvent(kMCMGR_Core0, kMCMGR_RemoteCoreUpEvent, 0);
+    status = mcmgr_trigger_event_internal(kMCMGR_Core0, kMCMGR_RemoteCoreUpEvent, 0U, false);
 #endif
 
     mcmgr_imu_remote_active_rel();
@@ -157,10 +154,12 @@ mcmgr_status_t mcmgr_get_core_property_internal(mcmgr_core_t coreNum,
     return kStatus_MCMGR_NotImplemented;
 }
 
-mcmgr_status_t mcmgr_trigger_event_internal(mcmgr_core_t coreNum, uint32_t remoteData, bool forcedWrite)
+mcmgr_status_t mcmgr_trigger_event_internal(mcmgr_core_t coreNum, mcmgr_event_type_t type, uint16_t eventData, bool forcedWrite)
 {
     (void)coreNum; /* Unused. IMU_LINK is making selections what core to trigger */
     (void)forcedWrite;
+
+    uint32_t remoteData = (((uint32_t)type) << 16U) | (uint32_t)eventData;
 
     int32_t ret;
     mcmgr_status_t status = kStatus_MCMGR_Error;