+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added initialization guard to public MCMGR API functions. All API functions that require a
+  fully initialized MCMGR instance now return `kStatus_MCMGR_NotReady` if `MCMGR_Init()` has
+  not been called successfully beforehand.
+
 ## [v5.2.0]
 
 ### Changed